# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EgliseHub API",
        "description": "Plateforme web multi-ministères chrétiens — République centrafricaine",
        "contact": {
            "name": "EgliseHub Support",
            "email": "admin@eglisehub.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local développement"
        },
        {
            "url": "https://api.eglisehub.org/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/ministry/worship-schedules": {
            "get": {
                "tags": [
                    "Worship Schedules"
                ],
                "summary": "Liste des horaires de cultes (Admin)",
                "description": "Récupère tous les horaires de cultes du ministère authentifié",
                "operationId": "getWorshipSchedules",
                "parameters": [
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Filtrer uniquement les horaires actifs",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des horaires",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WorshipSchedule"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Worship Schedules"
                ],
                "summary": "Créer un horaire de culte",
                "description": "Ajoute un nouvel horaire de culte pour le ministère",
                "operationId": "storeWorshipSchedule",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "jour",
                                    "heure_debut",
                                    "heure_fin"
                                ],
                                "properties": {
                                    "jour": {
                                        "type": "string",
                                        "enum": [
                                            "monday",
                                            "tuesday",
                                            "wednesday",
                                            "thursday",
                                            "friday",
                                            "saturday",
                                            "sunday"
                                        ],
                                        "example": "sunday"
                                    },
                                    "heure_debut": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "heure_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "12:00"
                                    },
                                    "is_highlight": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Culte de louange",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "ordre": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "ministere_id": {
                                        "description": "Uniquement pour super admin",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horaire créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Horaire ajouté avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WorshipSchedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "properties": {
                                                "jour": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "heure_debut": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "heure_fin": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/public/worship-schedules": {
            "get": {
                "tags": [
                    "Worship Schedules"
                ],
                "summary": "Liste publique des horaires de cultes",
                "description": "Récupère les horaires de cultes publics d'un ministère",
                "operationId": "getPublicWorshipSchedules",
                "parameters": [
                    {
                        "name": "subdomain",
                        "in": "query",
                        "description": "Sous-domaine du ministère",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "crc"
                        }
                    },
                    {
                        "name": "X-Subdomain",
                        "in": "header",
                        "description": "Sous-domaine du ministère (header)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "crc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des horaires publics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WorshipSchedule"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/ministry/worship-schedules/{id}/toggle-active": {
            "patch": {
                "tags": [
                    "Worship Schedules"
                ],
                "summary": "Activer/Désactiver un horaire",
                "description": "Bascule le statut actif/inactif d'un horaire",
                "operationId": "toggleWorshipScheduleActive",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'horaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Statut modifié."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WorshipSchedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Horaire non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "WorshipSchedule": {
                "required": [
                    "id",
                    "ministere_id",
                    "jour",
                    "heure_debut",
                    "heure_fin"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ministere_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "jour": {
                        "type": "string",
                        "enum": [
                            "monday",
                            "tuesday",
                            "wednesday",
                            "thursday",
                            "friday",
                            "saturday",
                            "sunday"
                        ],
                        "example": "sunday"
                    },
                    "heure_debut": {
                        "type": "string",
                        "format": "time",
                        "example": "09:00"
                    },
                    "heure_fin": {
                        "type": "string",
                        "format": "time",
                        "example": "12:00"
                    },
                    "is_highlight": {
                        "type": "boolean",
                        "example": false
                    },
                    "note": {
                        "type": "string",
                        "example": "Culte de louange",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "ordre": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}